+++ packages/database/test/database-sandbox.test.ts
@@ -1,12 +1,9 @@
-import { InMemoryTracer } from "@fwl/tracing";
-
 import * as database from "../index";
 
 let db: database.DatabaseQueryRunner;
-let tracer: InMemoryTracer;
-beforeAll(async () => {
-  tracer = new InMemoryTracer();
-  db = database.connect(tracer, {
+
+beforeEach(async () => {
+  db = await database.connectInSandbox({
     username: process.env.DATABASE_SQL_USERNAME || "fwl_db",
     host: process.env.DATABASE_SQL_HOST || "localhost",
     password: process.env.DATABASE_SQL_PASSWORD || "fwl_db",
@@ -17,13 +14,8 @@ beforeAll(async () => {
     "CREATE TABLE IF NOT EXISTS fwl (id UUID PRIMARY KEY, name varchar(32) NOT NULL)",
   );
 });
-beforeEach(async () => {
-  await db.query("TRUNCATE fwl");
-  tracer.spans = [];
-});
 
-afterAll(async () => {
-  await db.query("DROP TABLE fwl");
+afterEach(async () => {
   await db.close();
 });
 
@@ -75,7 +67,7 @@ describe("transactions", () => {
     }
   });
 
-  test("we should be able to manually rollback a transaction", async () => {
+  test("we should be able to manually rollback a transaction by throwing an error", async () => {
     expect.assertions(3);
     try {
       await db.transaction(async (client) => {
@@ -96,7 +88,7 @@ describe("transactions", () => {
     expect(rows.length).toBe(0);
   });
 
-  test("we should be able to manually rollback a transaction using the rollback function", async () => {
+  test.only("we should be able to manually rollback a transaction using the rollback function", async () => {
     expect.assertions(1);
     await db.transaction(async (client) => {
       await client.query("INSERT INTO fwl (id, name) VALUES ($1, $2)", [
@@ -105,6 +97,29 @@ describe("transactions", () => {
       ]);
       await client.rollback();
     });
+
+    const { rows } = await db.query("SELECT * FROM fwl WHERE name = $1", [
+      "in-transaction",
+    ]);
+    expect(rows.length).toBe(0);
+  });
+
+  test("it should return a TransactionError if we try to manually rollback a transaction by querying a ROLLBACK", async () => {
+    expect.assertions(3);
+    try {
+      await db.transaction(async (client) => {
+        await client.query("INSERT INTO fwl (id, name) VALUES ($1, $2)", [
+          "10f9a111-bf5c-4e73-96ac-5de87d962929",
+          "in-transaction",
+        ]);
+        await client.query("ROLLBACK");
+      });
+    } catch (error) {
+      expect(error).toBeInstanceOf(database.TransactionError);
+      expect(error.message).toBe(
+        "Can't query ROLLBACK inside of a `transaction` function, please throw an error or use transactionClient.rollback()",
+      );
+    }
 
     const { rows } = await db.query("SELECT * FROM fwl WHERE name = $1", [
       "in-transaction",
@@ -150,29 +165,6 @@ describe("transactions", () => {
     }
   });
 
-  test("it should return a TransactionError if we try to manually rollback a transaction by querying a ROLLBACK", async () => {
-    expect.assertions(3);
-    try {
-      await db.transaction(async (client) => {
-        await client.query("INSERT INTO fwl (id, name) VALUES ($1, $2)", [
-          "10f9a111-bf5c-4e73-96ac-5de87d962929",
-          "in-transaction",
-        ]);
-        await client.query("ROLLBACK");
-      });
-    } catch (error) {
-      expect(error).toBeInstanceOf(database.TransactionError);
-      expect(error.message).toBe(
-        "Can't query ROLLBACK inside of a `transaction` function, please throw an error or use transactionClient.rollback()",
-      );
-    }
-
-    const { rows } = await db.query("SELECT * FROM fwl WHERE name = $1", [
-      "in-transaction",
-    ]);
-    expect(rows.length).toBe(0);
-  });
-
   test("Should throw a DuplicateEntryError ", async () => {
     try {
       await db.query("INSERT INTO fwl (id, name) VALUES($1, $2)", [
@@ -197,55 +189,5 @@ describe("transactions", () => {
     } catch (error) {
       expect(error).toBeInstanceOf(database.BadUUIDError);
     }
-  });
-});
-
-describe("database tracing", () => {
-  test("The query should generate a corresponding span with the right name and attributes", async () => {
-    expect.assertions(3);
-    await db.query("INSERT INTO fwl (id, name) VALUES ($1, $2)", [
-      "74fbf638-6241-42bd-b257-b9a3dd24feb6",
-      "test",
-    ]);
-    const spans = tracer.searchSpanByName("db-query");
-    expect(spans.length).toBe(1);
-    const span = spans[0];
-    expect(span.name).toBe("db-query");
-    expect(span.attributes["db.statement"]).toBe(
-      "INSERT INTO fwl (id, name) VALUES ($1, $2)",
-    );
-  });
-
-  test("The transaction and associated query should generate the corresponding span with the right name and attributes", async () => {
-    expect.assertions(5);
-    await db.transaction(async (client) => {
-      await client.query("INSERT INTO fwl (id, name) VALUES ($1, $2)", [
-        "10f9a111-bf5c-4e73-96ac-5de87d962929",
-        "in-transaction",
-      ]);
-    });
-    const transactionSpan = tracer.searchSpanByName("db-transaction")[0];
-    expect(transactionSpan).toBeDefined();
-    expect(transactionSpan.name).toBe("db-transaction");
-    const querySpan = tracer.searchSpanByName("db-query")[0];
-    expect(querySpan).toBeDefined();
-    expect(querySpan.name).toBe("db-query");
-    expect(querySpan.attributes["db.statement"]).toBe(
-      "INSERT INTO fwl (id, name) VALUES ($1, $2)",
-    );
-  });
-
-  test("Closing the connection should generate the related span", async () => {
-    const testDb = database.connect(tracer, {
-      username: process.env.DATABASE_SQL_USERNAME || "fwl_db",
-      host: process.env.DATABASE_SQL_HOST || "localhost",
-      password: process.env.DATABASE_SQL_PASSWORD || "fwl_db",
-      database: process.env.DATABASE_SQL_DATABASE || "fwl_db",
-      port: parseFloat(process.env.DATABASE_SQL_PORT) || 5432,
-    });
-    await testDb.close();
-    const closeSpan = tracer.searchSpanByName("db-close")[0];
-    expect(closeSpan).toBeDefined();
-    expect(closeSpan.name).toBe("db-close");
   });
 });